Change miles from Int to Float
Allows mile tracking to be more granular and accurate

+++ Android Project/app/src/main/java/com/xyzian/cyclemech/SharedPreferencesManager.kt
@@ -11,14 +11,14 @@ class SharedPreferencesManager(context: Context){
         private const val MILES_KEY = "total_miles"
     }
 
-    fun setMiles(miles: Int) {
+    fun setMiles(miles: Float) {
         with(sharedPreferences.edit()) {
-            putInt(MILES_KEY, miles)
+            putFloat(MILES_KEY, miles)
             apply()
         }
     }
 
-    fun getMiles(): Int{
-        return sharedPreferences.getInt(MILES_KEY, 0)
+    fun getMiles(): Float{
+        return sharedPreferences.getFloat(MILES_KEY, 0.0F)
     }
 }
